+++ test/cbor/src/cbor_typed_array_tests.cpp
@@ -50,9 +50,9 @@ struct my_cbor_visitor : public default_json_visitor
     std::vector<double> v;
 private:
     JSONCONS_VISITOR_RET_TYP visit_typed_array(const span<const double>& data,  
-                        semantic_tag,
-                        const ser_context&,
-                        std::error_code&) override
+        semantic_tag,
+        const ser_context&,
+        std::error_code&) override
     {
         //std::cout << "visit_typed_array size: " << data.size() << "\n";
         v = std::vector<double>(data.begin(),data.end());
@@ -78,13 +78,21 @@ TEST_CASE("cbor Typed Array cursor tests")
         CHECK(staj_events::begin_array == cursor.current().event_type());
         CHECK(cursor.is_typed_array());
 
+        std::vector<double> v;
+        cursor.read_typed_array(v);
+        for (auto item : v)
+        {
+            std::cout << item << "\n";
+        }
+
         my_cbor_visitor visitor;
         cursor.read_to(visitor);
-        //for (auto item : visitor.v)
-        //{
-        //    std::cout << item << "\n";
-        //}
+        for (auto item : visitor.v)
+        {
+            std::cout << item << "\n";
+        }
     }
+#if 0 // test
     SECTION("Tag 86, float64, little endian, read_typed_array")
     {
         //std::cout << "CBOR cursor Typed Array Tag 86, float64, little endian" << '\n';
@@ -107,7 +115,10 @@ TEST_CASE("cbor Typed Array cursor tests")
         CHECK( -1.79769e+308 == Approx(v1[0]));
         CHECK(1.79769e+308 == Approx(v1[1]));
     }
+#endif // test
 }
+
+#if 0
 
 TEST_CASE("cbor Typed Array tests")
 {
@@ -1164,3 +1175,5 @@ TEST_CASE("multi-dim regular array and typed array")
         REQUIRE(expected == result);
     }
 }
+
+#endif // test